sistema/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-simple-history==3.10.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
idna==3.11
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg-pool==3.2.7
psycopg2-binary==2.9.11
PyJWT==2.10.1
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
